+++ packages/react/src/exampleTypes/Geo.js
@@ -63,7 +63,7 @@ let GeoComponent = injectTreeNode(
             <AutoComplete
               cacheOptions
               escapeClearsValue
-              defaultValue={node.selectedOption}
+              defaultInputValue={node.location}
               placeholder={placeholder}
               noOptionsMessage={() => ''}
               styles={customStyles}
@@ -72,14 +72,13 @@ let GeoComponent = injectTreeNode(
                 const inputValue = newValue.replace(/[^a-zA-Z0-9\s]+/g, '')
                 return inputValue
               }}
-              onChange={async selectedOption => {
-                let data = await GeoCodeLocation(selectedOption.value)
-                if (data && data.latitude && data.longitude && data.location) {
+              onChange={async ({ label, value }) => {
+                let data = await GeoCodeLocation(value)
+                if (data && data.latitude && data.longitude) {
                   tree.mutate(node.path, {
-                    selectedOption,
                     latitude: data.latitude,
                     longitude: data.longitude,
-                    location: data.location,
+                    location: label,
                   })
                 }
               }}